requirements.txt
pyautogen
chromadb
fastapi[standard]
pymongo